[Socket] Move SocketAsyncResult out of Socket_2_1.cs
