+++ mcs/class/System/System.dll.sources
@@ -536,6 +536,7 @@ System.Net.Sockets/Socket.cs
 System.Net.Sockets/Socket_2_1.cs
 System.Net.Sockets/SocketAsyncEventArgs.cs
 System.Net.Sockets/SocketAsyncOperation.cs
+System.Net.Sockets/SocketAsyncResult.cs
 System.Net.Sockets/SocketError.cs
 System.Net.Sockets/SocketException.cs
 System.Net.Sockets/SocketFlags.cs
